Support transaction notification

+++ src/main/kotlin/com/ojicoin/cookiepang/util/NotificationMessageUtils.kt
@@ -10,7 +10,7 @@ class NotificationMessageUtils {
         return "익명의 누군가가 ${cookieTitle}라는 질문을 요청했습니다."
     }
 
-    fun getTransactionMessage(nickname: String, cookieTitle: String, hammerCount: Double): String {
+    fun getTransactionMessage(nickname: String, cookieTitle: String, hammerCount: Long): String {
         // TODO consider supporting other language
 
         return "${nickname}님이 ${cookieTitle}를 망치 ${hammerCount}톤으로 구매했습니다."
